Add 'prompt' arg to import rsa privatekey function
Add an optional boolean 'prompt' argument to
interface.import_rsa_privatekey_from_file and change the behavior
of the function like so:

If password is passed use passed password for decryption.
If prompt is True use entered password for decryption.
If no password is passed or entered, or if the entered password
is an empty string, omit decryption.
Passing and prompting for a passowrd is not possible.

See code comments or secure-systems-lab/securesystemslib#122 for
more details.

This commit also adopts the unit tests accordingly.