FIX: gpg/rsa: fix #171 signature length bug
It appears the specification for PKCSv1.5 doesn't say how long should a
signature should be (i.e., should we zero-pad the signature or should we
use the minimum number of bits required). In this case, signatures
generated using gnupg would sometimes be 8*n bits shorter than the
public key length, which would result in signatures that fail to verify
due to missing leading zeroes (the leading zeroes existed at the bit
level however).

Zero-left-pad signature objects that do not match the length of the exponent
`n` on a gpg rsa public key object before verification.