Update keyid handling in parse_signature_packet
Update parse_signature_packet to only return the optional
short_keyid if it was parsed from the signature packet to (better)
match in_toto.gpg.formats.SIGNATURE_SCHEMA.

Note that the returned signature still might fail the schema check,
due to an empty keyid, but it is the responsibility of the caller
to adjust accordingly.

This commit also updates comments and calling functions.