Replace default-key with local-user in gpg sign
Pass keyids to gpg signing command using the --local-user
option instead of the --default-key option for consistent
handling of expired keys across gpg versions.

Testing the following command
```
echo bla | gpg --homedir . --detach-sign --default-key 2C199C7C
```
on "gpg (GnuPG/MacGPG2) 2.2.10" and "gpg (GnuPG) 1.4.18", only the
latter fails when trying to sign with an expired key, with the
following message on stderr:
  gpg: no default secret key: unusable secret key
  gpg: signing failed: unusable secret key

The newer version ignores the passed keyid and sign with the
keyring's default keyid, with the following message on stderr:
  gpg: Warning: not using '2C199C7C' as default key: No secret key
  gpg: all values passed to '--default-key' ignored

With --local-user signing fails on both versions.