Read fixed chunked size in `run_duplicate_streams`
This fixes a potential subversion of the timeout parameter
in `process.run_duplicate_streams`.

That is, if the child process writes to much and too fast to its
standard streams, the parent process will take too long (or even
indefinitely) to read and then write the contents to its own
standard streams, before it e.g. checks whether the process should
timeout.

This change adds a size parameter to the relevant read functions
to prevent above.