Unpin python-dateutil dependency
There is no reason to strictly pin python-dateutil to 2.8.0. On
the contrary, pinning dependency is prone to introduce dependency
conflicts (especially in downstream releases).

This was most likely a copy-paste mistake related to the practice
of pinning dependencies in dev-requirements.txt.

This is a quick fix for python-dateutil. A more comprehensive
dependency handling revision may be performed with
https://github.com/secure-systems-lab/securesystemslib/issues/182