Configure basic securesystemslib top-level logger
As pointed out by @FlorianVeaux in #210, securesystemslib does not
correctly initialize its loggers. As a consequence the logging
module logs a one-off "No handler could be found ..." warning when
used for the first time (only on Python 2).

The logging module then calls basicConfig, which attaches a
basic StreamHandler to the root logger, to which securesystemslib's
log messages are propagated.

This commit adds the configuration of a simple securesystemslib
top-level logger, with an attached StreamHandler and log level
WARNING, to which all other securesystemslib loggers default.

This removes the one-off warning, while providing a similar logging
behavior as before and clear instructions (see added comments) on
how to customize (format, silence, etc...) securesystemslib's logging.