Support nistp384 in ecdsa scheme and key formats
Support nistp384 in addition to nistp256 in ECDSA signing scheme
and key formats, required for PEP458.

This commit also removes the related but obsolete
REQUIRED_LIBRARIES_SCHEMA, which was only used in the
long-gone check_crypto_libraries tuf function.

Note: This commit tries to blend in with the current sslib design.
I future work we should:
- define securesystemslib-wide constants instead of hardcoding
  strings over and over again (see item 3 in #183)
- re-think "key type" vs. "signature scheme"