Restore PATH_SCHEMA to be AnyString
In one of the latest commits AnyNonemptyString was introduced
and with it, PATH_SCHEMA was changed to AnyNonemptyString from
AnyString. This led to unit tests failures in TUF.
See: https://github.com/theupdateframework/tuf/issues/1059

One could argue what is the right approach: to consider
that PATH_SCHEMA is every non-empty string or every string.
I think this is not of much importance, but for TUF and
possibly for other securesystemslib adopters, this leads to tests
failures. The reason is that there are multiple functions
with arguments which should conformant against the PATH_SCHEMA
but have default values of ''.

See :
- "prefix" argument in load_repository function
- "location_in_repository" argument in create_new_project function
both of which are in tuf/developer_tool.py

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>