include keyid in public ed25519/ecdsa key files
In the past we did not include the keyid in public ed25519/ecdsa key files. This commit changes this behavior and introduces a new argument to the `format_keyval_to_metadata` function. The argument is optional,
so this should not have any affect on using `format_keyval_to_metadata` for keyid generation.