Add and adopt encryption password interface helper
Add helper for key pair generation interface functions to use the
passed encryption password, or to prompt for a password based on a
passed 'prompt' boolean, or to not encrypt the key.

This helper aggregates and replaces repetitive code in
'interface.generate_and_write_{rsa, ed25519, ecdsa}_keypair'
functions, to make the password handling consistent accross these
functions and with private key import functions.

The commit adopts the helper for rsa and ed25519 generation
functions, adding an additional optional 'prompt' parameter.  The
ecdsa function requires more invasive changes (see subsequent
commit).

**Change of behavior**:
- Passing None as 'password' no longer opens a prompt, but just not
  encrypts the key. To open a prompt the new boolean kwarg 'prompt'
  must be used.
- Passing an empty password no longer writes the key unencrypted
  but raises a ValueError instead.

The goal of these changes is to require the caller to express the
encryption desire more explicitly.

Also see test changes for change of behavior.