gpg: Remove a circular dependency
There is a circular dependency in
  gpg.constants -> gpg.rsa -> gpg.util -> gpg.constants

This is not a problem on python3 or python2 when the imports are done
with "import securesystemslib.<mod>" style. However with the
"from securesystemslib import <mod>" style python2 decides this is a
ImportError.

Remove the circular dependency by moving the module variables from
constants to another file.