Optionally pad OpenPGP EdDSA signature parameters
Left-zero-pad 'r' and 's' values that are shorter than required by
RFC 8032 (5.1.6.), to make up for omitted leading zeros in RFC 4880
(3.2.) MPIs. This is especially important for 's', which is
little-endian.

Note that GnuPG seems to correctly parse EdDSA signatures with MPIs
that omit leading zeros prior to validation. But in order to
validate these signatures outside the context of RFC 4880, e.g.
with pyca/cryptography's OpenSSL backend, we need to guarantee the
correct length of the parsed signatures ourselves.