Add Tor project copyright notice to two files
formats.py and schema.py include code that is derivative of Thandy
projects source code. The code has now been licensed as MIT+Apache2.0
(and we're using it as MIT in Securesystemslib). To fulfill the license
requirements, let's make sure the copyright notices are correct.

Add Tor project copyright notice, and while we're at it add the other
relevant copyright notices (based on the history of the files in
question).

See also https://github.com/theupdateframework/tuf/issues/1491

Fixes #365