Add sign and verify methods to DSSE envelope
sign and verify methods are DSSE protocols used to create and verify
DSSE signatures. sign uses a Signer to sign the payload and create
DSSE a signature and verify method uses a Key List to verify all
DSSE signatures with the payload present in the envelope.

Veirfy function updated with latest Key specification. Now
try-catch the key.verify_signature instead of checking condition.

Signed-off-by: Pradyumna Krishna <git@onpy.in>