Remove colorama optional dependency
The coloroma library was used to optionally colorize the private
key file path in a password prompt, for encryption or decryption.

Given that the feature is in a rarely used code path -- interactive
calls are unlikely to be used outside of demo contexts -- and its
merit it generally debatable, our broad goal to minimize
dependencies weighs stronger.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>