Strip public key pem value
This commit adds a call to strip the public key pem value in
import_ecdsakey_from_private_pem.

The motivation for this that a later call to
import_ecdsakey_from_public_pem will otherwise cause an incorrect keyid
to be generated. This is because the call to `extract_pem` will strip
any new line character from the public pem, which will then be used to
generate the keyid, but this will be a different value than the keyid
generated earlier due to the removal of the new line character.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>