storage: convert from abc to typing.Protocol
The abstract base class pattern is useful when providing some shared
functionality for derivatives of the class. When the base class exists
only to provide an interface for implementers typing.Protocol, which
was introduced in Python 3.8, is a better fit.

Because Protocol support was added in Python 3.8, but we still support
Python 3.7 or newer, we need to use the typing-extensions module.
We require this dependency only for Python versions less than 3.8.

Signed-off-by: Joshua Lock <jlock@vmware.com>