signer: open session in sign method
Asking callers to handle the PyKCS11 session was a crutch to punt
the not quite straight-forward decision of how to identify the
correct token reader, but is no nice UX.

This commit moves session handling to the signer and just uses
the first reader slot that has a token.

This commit, move the session to the signer also requires loading
the pkcs11 library in the hsm module, and to take a login pin
(as secrets handler).

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>