signer: take pin not secrets handler in HSMSigner
When storing the secrets handler function as instance variable,
invocation will automatically pass self, but the secrets handler
only takes one argument.

There might be a workaround, but the easier solution is to just
call the secrets handler in from_priv_key_uri and pass the pin
to the constructor.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>