signer: standardize HSMSigner.import_() with GCPSigner
There is no signature to conform to here but since the methods do work
in a similar way, stanrdaize as much as possible:
 * same name
 * same return value

Returning the private key uri is not super useful for HSM at this point
(as the URI is a static string) but will be if e.g. slot is made user
configurable.