Use stdlib subprocess.run for gpg processes
This commit replaces calls to the custom process.run function, in
favor of the stdlib subprocess.run.

The former is a very thin wrapper around the latter, which was more
useful when we still needed Py2/3 case-handling. Which is not the
case anymore.

NOTE: This commit replicates the prior behavior, most notably the
following subprocess.run arguments:
- *popenargs: we still use shlex.split to turn the full command
              string into a list
- check

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>