Remove `securesystemslib.process` module
This module was originally developed in in-toto and transferred to
securesystemslib in #174, primarily as Py2/Py3-agnostic wrapper
around stdlib's `subprocess.run`, to run the `gpg` command. In #502
we switched to using `subprocess.run` directly.

Another wrapper around `subprocess.run`, provided by the module,
allows capturing standard streams AND write them to a terminal. It
was developed as specific `in-toto-run` feature and does not need
to be public API in sslib. in-toto/in-toto#544 moves the function
back to in-toto.

closes #345,
transfers #337 (to in-toto)

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>