Remove serialization module
Serialization api is removed from securesystemslib with this
commit because of its complexity and don't having a use case
for a standalone securesystemslib Envelope.

This commit removes `_default_deserializer`, `_default_serializer`
and `get_payload` methods of Envelope with securesystemslib but
keeps `from_dict` and `to_dict` methods to be consistent with
other metadata related classes, e.g. signer.

Signed-off-by: Pradyumna Krishna <git@onpy.in>