tests: rm HSM availability assertion in tox.ini
Currently, tox asserts that PYKCS11LIB env var is set to not
silently skip tests on SoftHSM (virtual HSM).

This is is unnecessary for CI, where we control the env var, and
also hard to figure out locally, where we can't expect everyone
to setup SoftHSM.

This patch removes the assertion and adds a note to ci.yml to point
out the consequences of unsetting the env var.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>