signer: assert keyid match GPGSigner import_, sign
Supporting subkeys in a GPGKey and considering them for
verification adds, for no benefit, a PKI hierarchy to tuf/in-toto,
which already have their own PKI hierarchies. In addition it makes
the verification (and delegation) code more complex and error
prone.

This commit drops subkey support, by the following two changes:

- import_: require exact match between passed keyid, and one of
  the keys in the bundle returned by gpg, and return a GPGKey
  only for that key, w/o subkeys

- sign: require exact match between keyid on attached public key
  and keyid on the signature returned by gpg

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>