signer: make sigstore signature API compatible
The Signature.signature field is defined as hex string of the
signature byte data. But for sigstore we need to store a complex
signature bundle, which also contains signature byte data.

To stay compatible with the API, we store both, the signature byte
data from the bundle in the Signature.signature field, and the
whole bundle under unrecognized_fields.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>