ci: file issue if sigstore test fails
Unfortunately, we cannot run the sigstore test workflow on PR open
because it requires 'id-token: write'-permissions, which is not
available in that context.

To still detect if a code change breaks the sigstore test, we run it
after the fact when the PR gets merged, and submit an issue in
case.

This strategy is copied from test-kms.yml.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>