signer: move _get_keyid staticmethod to Signer
This method can be used prior to creating a Key instance, in order
to generate a default keyid, which is a required Key argument.
Since this currently mostly happens in the import_ methods of
different Signer implementations, it feels right to implement this
on the Signer base class. But other places would be okay too.

Note that the computed keyids are different from
``securesystemslib.keys._get_keyid()``, which includes a legacy
field when producing the hash.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>