signer: use new _get_keyid in gcp and hsm signers
The import_ method of GCPSigner and HSMSigner used to assign
a default keyid to the imported key, computed with the legacy
_get_keyid function, which does overly complicated and
intransparent mangling of the input data.

This commit uses a new simple _get_keyid helper on the Signer base
class.

NOTE: keyid computation is not public API, and verification of
existing signatures with existing keys is not affected. If this
does disrupt existing users, who expect a certain value of a given
key, I'd rather annoy them now, where the signer API is still
experimental.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>