SigstoreSigner: Verify that signer token and key match
This is now possible with sigstore 2.x.

The verification is in from_priv_key_uri(). It could be in
constructor but I was trying to avoid making the optional
imports even more complicated...

Note that there is a workaround for the fact that we do not know
the signing identity at signing time in the ambient case:
The signing certificate SAN is not the OIDC identity.