Remove migrate key script but keep link in README
In preparation for the removal of legacy key interfaces and formats
(#731), we provided a key file migration script in docs/ (#658).

The script itself uses the legacy interfaces and thus should be removed
from the current git tree together with them.

This patch removes the script, including tests and test data, but keeps
a tagged reference in the README, so that it can still be used with the
pre-legacy-removal version of securesystemslib.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>