Remove 3 stray global key type constants
These feel a bit lost in the package-level namespace and are also
only a subset of the key types supported in the signer API.

Let's exclude them from the 1.0.0 API, and think of a suitable place
when addressing #593.

The patch also refactors an internal usage of one of the constants.
Externally, they seem to be only imported (but unused) in in-toto, which
is prepared for breaking changes in securesystemslib.

Signed-off-by: Lukas Puehringer <lukas.puehringer@nyu.edu>