requirements: Pin aws less strictly
Most of our dependabot PRs are for AWS libraries that release new
versions constantly. We only pin the libraries for test result
consistency: it seems that not pinning the patch version is
harmless here and stops a lot of the dependabot spam.

Pin only minor.major version for AWS libs.

Signed-off-by: Jussi Kukkonen <jkukkonen@google.com>